Docker version of visualmetrics
This makes it easier to run visualmetrics to collect timing metrics
from a video of a page loading. This will help us for T115857 and
T114425 because we can record a video locally and get waterfall graphs
from Firefox and metrics from the video. Hopefully this can also
help us get some sane values for proxy browsers.

Bug: T115857
Change-Id: Ifef9067b0b5796bc9c88c22ea78876694902ea35